requirements.txt
opencv-python==4.5.5.64
numpy==1.21.5
ultralytics==8.0.0
facenet-pytorch==2.5.2
opencv-python
Pillow==8.4.0
uv